requirements.txt
html2text==2024.2.26
IMAPClient==3.0.1
python-dotenv==1.0.1
